Entity key mapping from path parts
