+++ src/main/kotlin/org/coner/snoozle/db/entity/EntityDefinition.kt
@@ -5,11 +5,6 @@ import org.coner.snoozle.db.LiteralRecordDefinition
 
 class EntityDefinition<E : Entity<K>, K : Key> : LiteralRecordDefinition<E, K>() {
 
-    var entityKeyParser: EntityKeyParser<K>? = null
-        private set
-
-    fun entityKeyParser(fn: EntityKeyParserContext<E, K>.() -> K) {
-        entityKeyParser = EntityKeyParser(fn)
-    }
+    var entityKey: (EntityKeyParser<E, K>.Context.() -> K)? = null
 
 }
