+++ src/test/kotlin/org/coner/snoozle/db/sample/SampleDatabase.kt
@@ -8,11 +8,11 @@ class SampleDatabase(root: Path) : Database(root) {
     override val types = registerTypes {
         entity<Widget, WidgetKey> {
             path = "widgets" / { id } + ".json"
-            entityKeyParser { WidgetKey(id = uuidAt(0)) }
+            entityKey = { WidgetKey(id = uuidAt(0)) }
         }
         entity<Subwidget, SubwidgetKey> {
             path = "widgets" / { widgetId } / "subwidgets" / { id } + ".json"
-            entityKeyParser { SubwidgetKey(widgetId = uuidAt(0), id = uuidAt(1)) }
+            entityKey = { SubwidgetKey(widgetId = uuidAt(0), id = uuidAt(1)) }
         }
         versionedEntity<Gadget, GadgetKey> {
             path = "gadgets" / { id } / version + ".json"
